feat: add macro for ryot assets
