+++ crates/ryot_assets/src/atlas.rs
@@ -1,8 +1,6 @@
 use bevy_asset::Handle;
-use bevy_asset_loader::asset_collection::AssetCollection;
-use bevy_ecs::prelude::Resource;
 use bevy_sprite::TextureAtlasLayout;
 
-pub trait AtlasLayoutsAsset: Resource + AssetCollection + Send + Sync + 'static {
+pub trait AtlasLayoutsAsset: crate::RyotAsset {
     fn atlas_layouts(&self) -> &Vec<Handle<TextureAtlasLayout>>;
 }
